refactor(tests): split syndication.test.ts by format (atom/rss/json/shared)
890 行の syndication.test.ts を describe ブロック単位で 4 ファイルに機械的に分割。
assertion・seed・logic は一切変更せず、ファイル構造のみ整理した。

- atom.test.ts (260 行): /feed.atom, /feeds/:id.atom, /feeds/category/*.atom, /feeds/author/*.atom, /feeds/lang/*.atom
- json.test.ts (262 行): /feed.json, /feeds/:id.json, /feeds/category/*.json, /feeds/author/*.json, /feeds/lang/*.json
- rss.test.ts  (388 行): /feed.xml,  /feeds/:id.xml,  /feeds/category/*.xml,  /feeds/author/*.xml,  /feeds/lang/*.xml
- shared.test.ts (199 行): empty feed edge cases, special characters, ISO 8601 validation

worker 実装側の syndication/ ディレクトリ構成に合わせた。

Closes #369
